+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.x.x
+### Changed
+- Removed default key bindings
+
 ## 1.1.2
 ### Fixed
 - Fixed release notes